# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "IFS-Tests",
  "version": "0.1.0"
 },
 "paths": {
  "/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Me"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/logout": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Logout",
    "operationId": "logout",
    "responses": {
     "204": {
      "description": "Successful Response"
     }
    }
   }
  },
  "/auth/invites/{token}": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Invite Info",
    "operationId": "invite_info",
    "parameters": [
     {
      "name": "token",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InviteInfo"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/register": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Register",
    "operationId": "register",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Me"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/resets/{token}": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Reset Info",
    "operationId": "reset_info",
    "parameters": [
     {
      "name": "token",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ResetInfo"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/reset": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Reset",
    "operationId": "reset",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResetIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/me": {
   "get": {
    "tags": [
     "me"
    ],
    "summary": "Me",
    "operationId": "me",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Me"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "me"
    ],
    "summary": "Update Me",
    "operationId": "update_me",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Me"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/me/password": {
   "post": {
    "tags": [
     "me"
    ],
    "summary": "Change Password",
    "operationId": "change_password",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PasswordChangeIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Users",
    "operationId": "users",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AdminUser"
         },
         "type": "array",
         "title": "Response Users"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users/{user_id}": {
   "patch": {
    "tags": [
     "admin"
    ],
    "summary": "Update User",
    "operationId": "update_user",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AdminUser"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users/{user_id}/reset-link": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Reset Link",
    "operationId": "reset_link",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ResetLink"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/users/{user_id}/revoke-sessions": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Revoke Sessions",
    "operationId": "revoke_sessions",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/invites": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Open Invites",
    "operationId": "open_invites",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/OpenInvite"
         },
         "type": "array",
         "title": "Response Open Invites"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create Invite",
    "operationId": "create_invite",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InviteIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InviteCreated"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/invites/{invite_id}": {
   "delete": {
    "tags": [
     "admin"
    ],
    "summary": "Revoke Invite",
    "operationId": "revoke_invite",
    "parameters": [
     {
      "name": "invite_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Invite Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/audit": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Audit Log",
    "operationId": "audit_log",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AuditEntry"
         },
         "title": "Response Audit Log"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "AdminUser": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "vertical": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "Management",
         "Mechanical",
         "Tractive System",
         "Electronics",
         "Driverless",
         "Business",
         "Board"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Vertical"
     },
     "role": {
      "type": "string",
      "enum": [
       "member",
       "reviewer",
       "admin"
      ],
      "title": "Role"
     },
     "status": {
      "type": "string",
      "enum": [
       "active",
       "alumni",
       "disabled"
      ],
      "title": "Status"
     },
     "leaderboard_opt_out": {
      "type": "boolean",
      "title": "Leaderboard Opt Out"
     },
     "last_seen": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Seen"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "locked_until": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Locked Until"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "display_name",
     "vertical",
     "role",
     "status",
     "leaderboard_opt_out",
     "last_seen",
     "created_at",
     "locked_until"
    ],
    "title": "AdminUser"
   },
   "AuditEntry": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "at": {
      "type": "string",
      "format": "date-time",
      "title": "At"
     },
     "actor_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actor Id"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "target": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Target"
     },
     "details": {
      "additionalProperties": true,
      "type": "object",
      "title": "Details"
     }
    },
    "type": "object",
    "required": [
     "id",
     "at",
     "actor_id",
     "action",
     "target",
     "details"
    ],
    "title": "AuditEntry"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InviteCreated": {
    "properties": {
     "url": {
      "type": "string",
      "title": "Url"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "url",
     "expires_at"
    ],
    "title": "InviteCreated"
   },
   "InviteIn": {
    "properties": {
     "role": {
      "type": "string",
      "enum": [
       "member",
       "reviewer",
       "admin"
      ],
      "title": "Role",
      "default": "member"
     },
     "vertical": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "Management",
         "Mechanical",
         "Tractive System",
         "Electronics",
         "Driverless",
         "Business",
         "Board"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Vertical"
     },
     "note": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 80
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "type": "object",
    "title": "InviteIn"
   },
   "InviteInfo": {
    "properties": {
     "role": {
      "type": "string",
      "enum": [
       "member",
       "reviewer",
       "admin"
      ],
      "title": "Role"
     },
     "vertical": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "Management",
         "Mechanical",
         "Tractive System",
         "Electronics",
         "Driverless",
         "Business",
         "Board"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Vertical"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "role",
     "vertical",
     "expires_at"
    ],
    "title": "InviteInfo"
   },
   "LoginIn": {
    "properties": {
     "email": {
      "type": "string",
      "maxLength": 254,
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 256,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "LoginIn"
   },
   "Me": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "vertical": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "Management",
         "Mechanical",
         "Tractive System",
         "Electronics",
         "Driverless",
         "Business",
         "Board"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Vertical"
     },
     "role": {
      "type": "string",
      "enum": [
       "member",
       "reviewer",
       "admin"
      ],
      "title": "Role"
     },
     "leaderboard_opt_out": {
      "type": "boolean",
      "title": "Leaderboard Opt Out"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "display_name",
     "vertical",
     "role",
     "leaderboard_opt_out"
    ],
    "title": "Me"
   },
   "OpenInvite": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "role": {
      "type": "string",
      "enum": [
       "member",
       "reviewer",
       "admin"
      ],
      "title": "Role"
     },
     "vertical": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "Management",
         "Mechanical",
         "Tractive System",
         "Electronics",
         "Driverless",
         "Business",
         "Board"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Vertical"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "role",
     "vertical",
     "note",
     "created_at",
     "expires_at"
    ],
    "title": "OpenInvite"
   },
   "PasswordChangeIn": {
    "properties": {
     "current_password": {
      "type": "string",
      "maxLength": 256,
      "title": "Current Password"
     },
     "new_password": {
      "type": "string",
      "maxLength": 256,
      "title": "New Password"
     }
    },
    "type": "object",
    "required": [
     "current_password",
     "new_password"
    ],
    "title": "PasswordChangeIn"
   },
   "ProfileIn": {
    "properties": {
     "display_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "vertical": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "Management",
         "Mechanical",
         "Tractive System",
         "Electronics",
         "Driverless",
         "Business",
         "Board"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Vertical"
     },
     "clear_vertical": {
      "type": "boolean",
      "title": "Clear Vertical",
      "default": false
     },
     "leaderboard_opt_out": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Leaderboard Opt Out"
     }
    },
    "type": "object",
    "title": "ProfileIn"
   },
   "RegisterIn": {
    "properties": {
     "token": {
      "type": "string",
      "maxLength": 128,
      "title": "Token"
     },
     "email": {
      "type": "string",
      "maxLength": 254,
      "title": "Email"
     },
     "display_name": {
      "type": "string",
      "maxLength": 64,
      "title": "Display Name"
     },
     "password": {
      "type": "string",
      "maxLength": 256,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "token",
     "email",
     "display_name",
     "password"
    ],
    "title": "RegisterIn"
   },
   "ResetIn": {
    "properties": {
     "token": {
      "type": "string",
      "maxLength": 128,
      "title": "Token"
     },
     "password": {
      "type": "string",
      "maxLength": 256,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "token",
     "password"
    ],
    "title": "ResetIn"
   },
   "ResetInfo": {
    "properties": {
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "expires_at"
    ],
    "title": "ResetInfo"
   },
   "ResetLink": {
    "properties": {
     "url": {
      "type": "string",
      "title": "Url"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "url",
     "expires_at"
    ],
    "title": "ResetLink"
   },
   "UserPatch": {
    "properties": {
     "role": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "member",
         "reviewer",
         "admin"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Role"
     },
     "status": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "active",
         "alumni",
         "disabled"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     }
    },
    "type": "object",
    "title": "UserPatch"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}